requirements.txt
Flask==1.1.1
pytest==5.0.1  
python-dotenv==0.10.3 
Flask-SQLAlchemy==2.4.0  
